Fix type int to int64_t for windows compilation

+++ LibCarla/source/carla/trafficmanager/BatchControlStage.cpp
@@ -49,7 +49,7 @@ namespace traffic_manager {
     if (data_frame != nullptr &&
         number_of_vehicles != (*data_frame.get()).size()) {
 
-      number_of_vehicles = static_cast<uint>((*data_frame.get()).size());
+      number_of_vehicles = static_cast<uint64_t>((*data_frame.get()).size());
       // Allocating array for command batching.
       commands = std::make_shared<std::vector<cr::Command>>(number_of_vehicles);
     }
